version: fix runtime determined comparison against zero
This is an alternate form to fixing the GTK_SOURCE_CHECK_VERSION() which
performed the comparison in the macro where types are not available.

Related #178

+++ gtksourceview/gtksourceversion.c
@@ -114,5 +114,5 @@ gtk_source_check_version (guint major,
                           guint minor,
                           guint micro)
 {
-	return GTK_SOURCE_CHECK_VERSION (major, minor, micro);
+	return GTK_SOURCE_CHECK_VERSION (major, (int)minor, micro);
 }
